daemon: move channel `jsonrpc_` methods to an individual module
